requirements.txt
google-analytics-admin
google-analytics-data
functions-framework
google-auth